+++ CHANGELOG.md
@@ -6,6 +6,7 @@ IMPROVEMENTS:
 
  * Added snode.service_address to get the IP address of a service easily
    without any logic in ERB files
+ * Improved consul-ui, added a [README.md](samples/consul-ui/README.md)
 
 ## 1.1.3 (May 17, 2018)